+++ src/main/java/com/api/apifinanceiro/repositories/FinanceiroRepository.java
@@ -1,6 +1,7 @@
 package com.api.apifinanceiro.repositories;
 
 import com.api.apifinanceiro.entities.models.FinanceiroModel;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,5 +9,6 @@ import java.util.UUID;
 
 @Repository
 public interface FinanceiroRepository extends JpaRepository<FinanceiroModel, UUID> {
+  Optional<FinanceiroModel> findByIdFuncionario(UUID idFuncionario);
 
 }
