--- src/main/java/com/api/apifinanceiro/utils/postmanFile/API_FINANCEIRO.postman_collection.json
@@ -1,143 +1,0 @@
-{
-	"info": {
-		"_postman_id": "d56ce585-3e42-42bb-a6ba-2665a85838fb",
-		"name": "API_FINANCEIRO",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "29578706"
-	},
-	"item": [
-		{
-			"name": "POST",
-			"item": [
-				{
-					"name": "Cadastro - COMPLETO VÁLIDO",
-					"request": {
-						"method": "POST",
-						"header": [],
-						"body": {
-							"mode": "raw",
-							"raw": "{\r\n    \"idFuncionario\": \"550e8400-e29b-41d4-a716-446655440000\",\r\n    \"cargo\": \"INSTRUTOR\",\r\n    \"dataAdmissao\": \"2023-09-29T14:30+03:00\",\r\n    \"salario\": 3500.00\r\n}",
-							"options": {
-								"raw": {
-									"language": "json"
-								}
-							}
-						},
-						"url": {
-							"raw": "http://localhost:8080/v1/financeiro",
-							"protocol": "http",
-							"host": [
-								"localhost"
-							],
-							"port": "8080",
-							"path": [
-								"v1",
-								"financeiro"
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Cadastro - INCOMPLETO VÁLIDO",
-					"request": {
-						"method": "POST",
-						"header": [],
-						"body": {
-							"mode": "raw",
-							"raw": "{\r\n    \"idFuncionario\": \"550e8400-e29b-41d4-a716-446655440000\",\r\n    \"cargo\": \"INSTRUTOR\",\r\n    \"dataAdmissao\": \"2023-09-29T14:30+03:00\",\r\n    \"salario\": 3500.00\r\n}",
-							"options": {
-								"raw": {
-									"language": "json"
-								}
-							}
-						},
-						"url": {
-							"raw": "http://localhost:8080/v1/financeiro",
-							"protocol": "http",
-							"host": [
-								"localhost"
-							],
-							"port": "8080",
-							"path": [
-								"v1",
-								"financeiro"
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Cadastro - CAMPO VAZIO",
-					"request": {
-						"method": "POST",
-						"header": [],
-						"body": {
-							"mode": "raw",
-							"raw": "{\r\n    \"idFuncionario\": \"550e8400-e29b-41d4-a716-446655440000\",\r\n    \"cargo\": \"INSTRUTOR\",\r\n    \"dataAdmissao\": \"2023-09-29T14:30+03:00\",\r\n    \"salario\": 3500.00\r\n}",
-							"options": {
-								"raw": {
-									"language": "json"
-								}
-							}
-						},
-						"url": {
-							"raw": "http://localhost:8080/v1/financeiro",
-							"protocol": "http",
-							"host": [
-								"localhost"
-							],
-							"port": "8080",
-							"path": [
-								"v1",
-								"financeiro"
-							]
-						}
-					},
-					"response": []
-				},
-				{
-					"name": "Cadastro - SEM UM CAMPO OBRIGATORIO",
-					"request": {
-						"method": "POST",
-						"header": [],
-						"body": {
-							"mode": "raw",
-							"raw": "{\r\n    \"idFuncionario\": \"550e8400-e29b-41d4-a716-446655440000\",\r\n    \"cargo\": \"INSTRUTOR\",\r\n    \"dataAdmissao\": \"2023-09-29T14:30+03:00\",\r\n    \"salario\": 3500.00\r\n}",
-							"options": {
-								"raw": {
-									"language": "json"
-								}
-							}
-						},
-						"url": {
-							"raw": "http://localhost:8080/v1/financeiro",
-							"protocol": "http",
-							"host": [
-								"localhost"
-							],
-							"port": "8080",
-							"path": [
-								"v1",
-								"financeiro"
-							]
-						}
-					},
-					"response": []
-				}
-			]
-		},
-		{
-			"name": "UPDATE",
-			"item": []
-		},
-		{
-			"name": "DELETE",
-			"item": []
-		},
-		{
-			"name": "GET",
-			"item": []
-		}
-	]
-}
