--- src/test/java/com/api/apifinanceiro/ApiFinanceiroApplicationTests.java
@@ -1,14 +1,0 @@
-package com.api.apifinanceiro;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiFinanceiroApplicationTests {
-
-	@Test
-	void contextLoads() {
-
-	}
-
-}
