--- src/test/java/com/api/apifinanceiro/FinanceiroControllerTest.java
@@ -1,18 +1,0 @@
-package com.api.apifinanceiro;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-@DisplayName("FinanceiroController - Testes")
-public class FinanceiroControllerTest {
-
-  @Test
-  @DisplayName("01 - Controller - Criar financeiro válido")
-  public void testarCriarFinanceiroValido() {
-  }
-
-  @Test
-  @DisplayName("02 - Controller - Criar financeiro inválido")
-  public void testarCriarFinanceiroInvalido() {
-  }
-}
